refactor(clean-architecture): Merge medit-plugins into app
Current Android Room logic is actually associated with the Android App itself. Android App also represents plugins layer of clean architecture.
