+++ app/build.gradle
@@ -1,8 +1,7 @@
-plugins {
-    id 'com.android.application'
-    id 'kotlin-android'
-    id 'org.jlleitschuh.gradle.ktlint'
-}
+apply plugin: 'com.android.application'
+apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-kapt'
+apply plugin: 'org.jlleitschuh.gradle.ktlint'
 
 android {
     compileSdkVersion 30
@@ -39,8 +38,12 @@ android {
 
 dependencies {
     final def markwon_version = '4.6.2'
+    final def room_version = "2.2.6"
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+
+    implementation project(":medit-domain")
+
     implementation 'androidx.core:core-ktx:1.3.2'
     implementation 'androidx.appcompat:appcompat:1.2.0'
     implementation 'com.google.android.material:material:1.3.0'
@@ -59,6 +62,17 @@ dependencies {
     implementation "io.noties.markwon:ext-latex:$markwon_version"
     implementation "io.noties.markwon:inline-parser:$markwon_version"
     /* ---------------- */
+
+    /* Android Room framework */
+    implementation "androidx.room:room-runtime:$room_version"
+    kapt "androidx.room:room-compiler:$room_version"
+
+    // optional - Kotlin Extensions and Coroutines support for Room
+    implementation "androidx.room:room-ktx:$room_version"
+
+    // optional - Test helpers
+    testImplementation "androidx.room:room-testing:$room_version"
+    /* ---------------------- */
     
     /* Test frameworks */
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
